requirements.txt
flask==2.3.3
markdown==3.4.4
python-dotenv==1.0.0
werkzeug==2.3.7